feat: adaptive tags icon
